requirements.txt
pandas
scikit-learn
openpyxl
newspaper3k   
